+++ fathom-archetype-standard/src/main/resources/archetype-resources/src/main/java/controllers/SecureApiController.java
@@ -61,7 +61,6 @@ public class SecureApiController extends Controller {
      *
      * @param id
      * @param account
-     * @return Reply
      */
     @GET("/{id: [0-9]+}")
     @Produces({Produces.JSON, Produces.XML})