+++ fathom-integration-test/src/main/java/controllers/SecureApiController.java
@@ -22,10 +22,11 @@ import fathom.metrics.Metered;
 import fathom.realm.Account;
 import fathom.rest.controller.Auth;
 import fathom.rest.controller.Controller;
-import fathom.rest.controller.Path;
 import fathom.rest.controller.GET;
+import fathom.rest.controller.Path;
 import fathom.rest.controller.Produces;
 import fathom.rest.security.aop.RequirePermission;
+import fathom.rest.swagger.Undocumented;
 import models.Item;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -34,6 +35,7 @@ import org.slf4j.LoggerFactory;
  * To be discoverable, a controller must be annotated with {@code @Path}
  */
 @Path("/secure/")
+@Undocumented
 public class SecureApiController extends Controller {
 
     private final Logger log = LoggerFactory.getLogger(SecureApiController.class);
@@ -59,7 +61,6 @@ public class SecureApiController extends Controller {
      *
      * @param id
      * @param account
-     * @return Reply
      */
     @GET("{id: [0-9]+}")
     @Produces({Produces.JSON, Produces.XML})