+++ fathom-rest-swagger/src/main/java/fathom/rest/swagger/Tag.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package fathom.rest.controller;
+package fathom.rest.swagger;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -21,13 +21,13 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * Specifies HTTP PUT method.
+ * Annotation for an controller Tag.
  *
  * @author James Moger
  */
-@Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
-@HttpMethod(HttpMethod.PUT)
-public @interface PUT {
-    String[] value() default {};
+@Target(ElementType.TYPE)
+public @interface Tag {
+    String name() default "";
+    String description();
 }