+++ fathom-rest-swagger/src/test/java/fathom/rest/swagger/SwaggerBuilderTest.java
@@ -1,0 +1,94 @@
+/*
+ * Copyright (C) 2015 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package fathom.rest.swagger;
+
+import com.google.inject.Guice;
+import com.google.inject.Injector;
+import fathom.conf.Settings;
+import fathom.rest.controller.Controller;
+import fathom.rest.controller.ControllerHandler;
+import fathom.rest.controller.DELETE;
+import fathom.rest.controller.GET;
+import fathom.rest.controller.HttpMethod;
+import fathom.rest.controller.PATCH;
+import fathom.rest.controller.POST;
+import fathom.rest.controller.PUT;
+import fathom.rest.controller.Path;
+import fathom.rest.controller.Produces;
+import org.junit.Test;
+import ro.pippo.core.route.DefaultRouter;
+import ro.pippo.core.route.Route;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import static org.junit.Assert.*;
+
+/**
+ * @author James Moger
+ */
+public class SwaggerBuilderTest {
+
+    @Test
+    public void testGenerate() throws Exception {
+        SwaggerBuilder sb = new SwaggerBuilder(new Settings(), new DefaultRouter());
+        String json = sb.generateJSON(getRoutes());
+        assertNotNull(json);
+    }
+
+    private List<Route> getRoutes() {
+        Injector injector = Guice.createInjector();
+
+        List<Route> routes = new ArrayList<>();
+        routes.add(route(injector, "GET", "/api/{id}", TestController.class));
+        routes.add(route(injector, "PUT", "/api/{id}", TestController.class));
+        routes.add(route(injector, "POST", "/api/{id}", TestController.class));
+        routes.add(route(injector, "PATCH", "/api/{id}", TestController.class));
+        routes.add(route(injector, "DELETE", "/api/{id}", TestController.class));
+        return routes;
+    }
+
+    private Route route(Injector injector, String method, String path, Class<? extends Controller> controllerClass) {
+        ControllerHandler handler = new ControllerHandler(injector, controllerClass, method.toLowerCase());
+        return new Route(path, method, handler);
+    }
+
+    public static class TestController extends Controller {
+
+        @Produces({Produces.JSON, Produces.XML})
+        public void get(int id) {
+        }
+
+        @Produces({Produces.JSON, Produces.XML})
+        public void put(int id) {
+        }
+
+        @Produces(Produces.HTML)
+        public void post(int id) {
+        }
+
+        @Produces({Produces.JSON, Produces.XML})
+        public void patch(int id) {
+        }
+
+        @Produces({Produces.JSON, Produces.XML})
+        public void delete(int id) {
+        }
+
+    }
+
+}